+++ src/app/components/game/game.component.ts
@@ -75,7 +75,7 @@ export class GameComponent implements OnInit, OnChanges {
     private initPosition: Board;
 
     // event listener functions
-    private _mouseEnterEventListener: () => void;
+    private _mouseEnterEventListener: (event: any) => void;
     private _mouseLeaveEventListener: () => void;
     private _mouseMoveEventListener: (event: any) => void;
     private _mouseDownEventListener: (event: any) => void;
@@ -108,9 +108,24 @@ export class GameComponent implements OnInit, OnChanges {
         this.matchingMoves = [];
         this.displayedMoveIndex = 0;
 
-        this._mouseEnterEventListener = () => {
+        this._mouseEnterEventListener = (event: any) => {
             // just a detector of when the mouse is over the canvas object
             this.cursor.mouseOverBoard = true;
+            // if mouse was down end that "drag drop event" if the mouse isn't
+            // still down
+            if (this.cursor.leftMouseIsDown) {
+                if (event.buttons === 0 || event.which !== 1) {
+                    // pause here, but this is where we put the "mouse back up"
+                    // left button
+                    this.cursor.leftMouseIsDown = false;
+                    this.cursor.dragging = false;
+                    this.cursor.draggedPieceIndex = -1;
+                    this.drawBoard();
+                    // this block is the same logic in mouseUpEvent but without
+                    // the action of doing a bunch of promotion / make move
+                    // checking
+                }
+            }
         };
         this._mouseLeaveEventListener = () => {
             // when mouse exits the canvas object
@@ -673,7 +688,10 @@ export class GameComponent implements OnInit, OnChanges {
             this.drawCoordinates();
         }
         this.boardCtx.globalAlpha = 1;
-        if (this.cursor.draggedPieceIndex !== -1) {
+        if (
+            this.cursor.draggedPieceIndex !== -1 &&
+            this.cursor.mouseOverBoard
+        ) {
             let x = this.cursor.currentMousePosition.x;
             let y = this.cursor.currentMousePosition.y;
             if (this.displayOptions.orientation === Color.Black) {